# Scripts: Only output esm for nodeEntries on request