--- a/code/addons/test/package.json
+++ b/code/addons/test/package.json
@@ -68,6 +68,7 @@
   "types": "dist/index.d.ts",
   "files": [
     "dist/**/*",
+    "!dist/dummy.*",
     "README.md",
     "*.mjs",
     "*.js",