Closes #

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

<!-- Briefly describe what your PR does -->

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  77.8 MB | 77.8 MB | 0 B | **-2.12** | 0% |
| initSize |  131 MB | 131 MB | 0 B | -0.34 | 0% |
| diffSize |  53 MB | 53 MB | 0 B | -0.33 | 0% |
| buildSize |  7.19 MB | 7.19 MB | 0 B | 1.08 | 0% |
| buildSbAddonsSize |  1.85 MB | 1.85 MB | 0 B | - | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.87 MB | 1.87 MB | 0 B | -0.82 | 0% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.91 MB | 3.91 MB | 0 B | -0.82 | 0% |
| buildPreviewSize |  3.28 MB | 3.28 MB | 0 B | 1.09 | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  11.8s | 6.7s | -5s -126ms | **-1.37** | 🔰-76.1% |
| generateTime |  22.5s | 19.1s | -3s -413ms | -1.15 | -17.9% |
| initTime |  15.1s | 12.8s | -2s -248ms | -1.1 | -17.5% |
| buildTime |  8.2s | 9.7s | 1.4s | 0.21 | 15.3% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5.2s | 4.8s | -383ms | -0.48 | -7.8% |
| devManagerResponsive |  3.9s | 3.6s | -343ms | -0.43 | -9.5% |
| devManagerHeaderVisible |  586ms | 597ms | 11ms | -0.48 | 1.8% |
| devManagerIndexVisible |  613ms | 662ms | 49ms | -0.07 | 7.4% |
| devStoryVisibleUncached |  2.1s | 2s | -153ms | -0.13 | -7.5% |
| devStoryVisible |  614ms | 661ms | 47ms | -0.15 | 7.1% |
| devAutodocsVisible |  549ms | 499ms | -50ms | -0.75 | -10% |
| devMDXVisible |  490ms | 523ms | 33ms | -0.44 | 6.3% |
| buildManagerHeaderVisible |  530ms | 540ms | 10ms | -0.81 | 1.9% |
| buildManagerIndexVisible |  621ms | 628ms | 7ms | -0.82 | 1.1% |
| buildStoryVisible |  519ms | 531ms | 12ms | -0.77 | 2.3% |
| buildAutodocsVisible |  415ms | 405ms | -10ms | **-1.53** | -2.5% |
| buildMDXVisible |  407ms | 427ms | 20ms | -0.93 | 4.7% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Based on the provided files and context, I'll create a concise summary of the pull request changes:

Optimized the bundling configuration for node entries in Storybook addons to allow selective ESM/CJS output formats, with specific implementation in the test addon.

- Modified `code/addons/test/package.json` to specify explicit ESM/CJS output for vitest-plugin files
- Updated `scripts/prepare/addon-bundle.ts` to support format-specific node entry bundling
- Added new `NodeEntry` type to handle format preferences per entry
- Maintained backward compatibility by treating string entries as CJS-only
- Separated build process into ESM and CJS builds based on entry preferences



<!-- /greptile_comment -->